added missing assets, upgraded 2d toolkit

+++ Assets/TK2DROOT/tk2d/Shaders/LitCutoutVertexColor.shader
@@ -15,13 +15,13 @@ Shader "tk2d/LitCutoutVertexColor"
 		CGPROGRAM
 		#pragma surface surf Lambert alphatest:_Cutoff
 		struct Input {
-			float2 uv_MainTex;
-			float4 color : COLOR;
+			half2 uv_MainTex;
+			fixed4 color : COLOR;
 		};
 		sampler2D _MainTex;
 		void surf(Input IN, inout SurfaceOutput o)
 		{
-			half4 mainColor = tex2D(_MainTex, IN.uv_MainTex) * IN.color;
+			fixed4 mainColor = tex2D(_MainTex, IN.uv_MainTex) * IN.color;
 			o.Albedo = mainColor.rgb;
 			o.Alpha = mainColor.a;
 		}
